Move doc routes to own file
